Refactoring and fix for filtering #172115886
