+++ app/javascript/components/supplemental_application/actions.js
@@ -1,25 +1,8 @@
 import apiService from '~/apiService'
 import Alerts from '~/components/Alerts'
-import { isEmpty, find, isEqual, every, reject, forEach, isPlainObject } from 'lodash'
+import { isEmpty, find, isEqual, every, reject } from 'lodash'
 import { convertCurrency } from '~/utils/form/validations'
-
-const filterChanged = (prev, current) => {
-  const changedFields = {}
-  forEach(current, (value, key) => {
-    if (!isEqual(prev[key], value)) {
-      if (isPlainObject(value)) {
-        const obj = filterChanged(prev[key], value)
-        if (!isEmpty(obj)) {
-          obj['id'] = value.id
-          changedFields[key] = obj
-        }
-      } else {
-        changedFields[key] = value
-      }
-    }
-  })
-  return changedFields
-}
+import { filterChanged } from '~utils/utils'
 
 export const updateApplication = async (application, prevApplication) => {
   const primaryApplicantContact = application.applicant && application.applicant.id
